use props
